fix: replace robotjs by nutjs (#2295) (#2302)
* fix: replace robotjs by nutjs (#2295)

* fix: do not create quick ask window on test
